refactor(api): reorganize service result classes into specific packages for clarity
